src_visualiser_open3d/requirements.txt
Flask
numpy
open3d
azure-storage-blob
Werkzeug
threaded